Added some code

+++ cpp-yellow/week-1/block_weight.cpp
@@ -6,15 +6,15 @@ using namespace std;
 int main()
 {
     int n;
-    int64_t r;
+    uint64_t r;
 
     cin >> n >> r;
 
-    int64_t weight = 0;
+    uint64_t weight = 0;
 
     for (int i = 0; i < n; i++)
     {
-        int w, h, d;
+        uint64_t w, h, d;
         cin >> w >> h >> d;
 
         weight += w * h * d * r;
